+++ src/ShareTravelSystem.Web/Controllers/TownController.cs
@@ -1,30 +1,31 @@
 ﻿namespace ShareTravelSystem.Web.Controllers
 {
-    using Microsoft.AspNetCore.Authorization;
-    using Microsoft.AspNetCore.Identity;
     using Microsoft.AspNetCore.Mvc;
     using ShareTravelSystem.Services.Contracts;
     using ShareTravelSystem.ViewModels;
     using ShareTravelSystem.ViewModels.Town;
-    using ShareTravelSystem.Web.Areas.Identity.Data;
     using System;
+    using ShareTravelSystem.Web.Infrastructure.Constants;
 
-    [Authorize(Roles = "Admin")]
-    public class TownController : Controller
+    public class TownController : BaseController
     {
+        private const string redirectToPath = "/Account/Login";
         private readonly ITownService townService;
-        private readonly UserManager<ShareTravelSystemUser> userManager;
 
-        public TownController(ITownService townService, UserManager<ShareTravelSystemUser> userManager)
+        public TownController(ITownService townService)
         {
             this.townService = townService;
-            this.userManager = userManager;
         }
 
         [HttpGet]
         public IActionResult All(int page, string search)
         {
-            int size = 10;
+            if (!RedirectIfNotInRole(Constants.AdminRole))
+            {
+                return Redirect(redirectToPath);
+            };
+
+            int size = Constants.TownsPerPage;
             TownPaginationViewModel towns = this.townService.GetAllTowns(size, page, search);
 
             return View(towns);
@@ -33,6 +34,11 @@
         [HttpGet]
         public IActionResult Create()
         {
+            if (!this.RedirectIfNotInRole(Constants.AdminRole))
+            {
+                return Redirect(redirectToPath);
+            };
+
             return View();
         }
 
@@ -40,6 +46,11 @@
         [ValidateAntiForgeryToken]
         public IActionResult Create(CrateTownViewModel model)
         {
+            if (!this.RedirectIfNotInRole(Constants.AdminRole))
+            {
+                return Redirect(redirectToPath);
+            };
+
             if (!ModelState.IsValid)
             {
                 return this.View(model);
@@ -63,7 +74,13 @@
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            if (!this.RedirectIfNotInRole(Constants.AdminRole))
+            {
+                return Redirect(redirectToPath);
+            };
+
             EditTownViewModel town;
+
             try
             {
                 town = this.townService.GetTownById(id);
@@ -75,6 +92,7 @@
                 return RedirectToAction(nameof(TownController.All));
 
             }
+
             return View(town);
         }
 
@@ -82,6 +100,11 @@
         [ValidateAntiForgeryToken]
         public IActionResult Edit(EditTownViewModel model)
         {
+            if (!this.RedirectIfNotInRole(Constants.AdminRole))
+            {
+                return Redirect(redirectToPath);
+            };
+
             if (!ModelState.IsValid)
             {
                 return this.View(model);
@@ -94,7 +117,6 @@
             catch (Exception e)
             {
                 this.ModelState.AddModelError("Name", e.Message);
-
                 return this.View(model);
             }
 
@@ -105,6 +127,11 @@
         [ValidateAntiForgeryToken]
         public IActionResult Delete(int id)
         {
+            if (!this.RedirectIfNotInRole(Constants.AdminRole))
+            {
+                return Redirect(redirectToPath);
+            }
+
             try
             {
                 this.townService.Delete(id);
@@ -112,9 +139,9 @@
             catch (Exception e)
             {
                 this.ModelState.AddModelError("Name", e.Message);
-
                 return RedirectToAction(nameof(TownController.All));
             }
+
             return RedirectToAction(nameof(TownController.All));
         }
     }
